requirements.txt
attrs==18.2.0